wedge-worker/requirements.txt
paho-mqtt
networkx
onnx==1.11
